fix: Error will be occurred with native popupmenu
Problem: An error occurs when displaying a preview. It's caused by a
difference between |pum_getpos()| and |pum#getpos()|. It is a type
mismatch of `scrollbar` field. The field is defined as |TRUE| in their
documentation. However, it is used as integer in View:_open()
function.

Solution: Check `scrollbar` field by |empty()| that it's vim |TRUE|.